Removed Statics, migrated methods Removed all methods from Statics, migrating to more appropriate files. Reformatted some sections of Program.cs.
