+++ DiscordInterop/SlashCommands.cs
@@ -6,7 +6,7 @@ namespace StarBot.DiscordInterop;
 internal class SlashCommands {
     public static async Task keySet(SocketSlashCommand command, DiscordSocketClient? client, Database data) {
 
-        if (Statics.userHasRole(client, command.GuildId, command.User.Id, Config.ADMIN_ROLE_ID)) {
+        if (UserManager.userHasRole(client, command.GuildId, command.User.Id, Config.ADMIN_ROLE_ID)) {
             var commandArgs = command.Data.Options.ToArray();
             string? key = commandArgs[0].Value.ToString();
             string? value = commandArgs[1].Value.ToString();
@@ -46,7 +46,7 @@ internal class SlashCommands {
         }
         string uiStatus = "";
         if (interested == 1) {
-            if (!Statics.userHasRole(client, command.GuildId, command.User.Id, Config.STARBOT_INTEREST_ROLE_ID)) {
+            if (!UserManager.userHasRole(client, command.GuildId, command.User.Id, Config.STARBOT_INTEREST_ROLE_ID)) {
                 uiStatus = "APPROVED";
                 await client.GetGuild((ulong)command.GuildId).GetUser(command.User.Id).AddRoleAsync(Config.STARBOT_INTEREST_ROLE_ID);
                 await command.RespondAsync($"Your interest in StarBot is appreciated. Your access to backend channels was **{uiStatus}**.\nThis action has been logged. Access to these channels may be revoked at any time for any reason.", ephemeral: true);
@@ -55,7 +55,7 @@ internal class SlashCommands {
                 return;
             }
         } else {
-            if (Statics.userHasRole(client, command.GuildId, command.User.Id, Config.STARBOT_INTEREST_ROLE_ID)) {
+            if (UserManager.userHasRole(client, command.GuildId, command.User.Id, Config.STARBOT_INTEREST_ROLE_ID)) {
                 uiStatus = "REMOVED";
                 await client.GetGuild((ulong)command.GuildId).GetUser(command.User.Id).RemoveRoleAsync(Config.STARBOT_INTEREST_ROLE_ID);
                 await command.RespondAsync($"Your access to backend channels was **{uiStatus}**.\nThis action has been logged.", ephemeral: true);
@@ -74,7 +74,7 @@ internal class SlashCommands {
     }
     public static async Task executeTask(SocketSlashCommand command, Scheduler scheduler, DiscordSocketClient client, Database data, MemoryCacheManager cache) {
         await command.DeferAsync(ephemeral: true);
-        if (!Statics.userHasRole(client, command.GuildId, command.User.Id, Config.ADMIN_ROLE_ID)) {
+        if (!UserManager.userHasRole(client, command.GuildId, command.User.Id, Config.ADMIN_ROLE_ID)) {
             await command.FollowupAsync("You do not have the required permissions to execute this command.", ephemeral: true);
             return;
         }
